修改资料,性别修改不生效bug. 没班级点击创建学校,提示为请先创建班级

+++ src/main/java/com/company/news/rest/UserinfoController.java
@@ -792,7 +792,10 @@ public class UserinfoController extends AbstractRESTController {
 		// 返回消息体
 		ResponseMessage responseMessage = RestUtil
 				.addResponseMessageForModelMap(model);
-		
+		if(!RightUtils.hasRight(SystemConstants.Group_uuid_wjkj, RightConstants.AD_user_del, request)){
+			responseMessage.setMessage("没有删除权限");
+			return "";
+		}
 		try {
 			
 			boolean flag = userinfoService.delete(request.getParameter("uuid"),
